@@ -33,57 +33,57 @@
   },
   "devDependencies": {
     "@sindresorhus/slugify": "1.1.2",
-    "@slack/webhook": "6.0.0",
+    "@slack/webhook": "6.1.0",
     "@types/benchmark": "2.1.1",
     "@types/glob": "7.2.0",
     "@types/graphviz": "0.0.34",
     "@types/node": "14.18.17",
     "@types/redis": "2.8.32",
     "@types/resolve": "1.20.2",
-    "@typescript-eslint/eslint-plugin": "5.9.0",
-    "@typescript-eslint/parser": "5.9.0",
+    "@typescript-eslint/eslint-plugin": "5.23.0",
+    "@typescript-eslint/parser": "5.23.0",
     "arg": "5.0.1",
     "batching-toposort": "1.2.0",
     "buffer": "6.0.3",
     "chalk": "4.1.2",
-    "chokidar": "3.5.2",
-    "esbuild": "0.13.14",
-    "esbuild-register": "3.3.1",
-    "eslint": "8.6.0",
-    "eslint-config-prettier": "8.3.0",
+    "chokidar": "3.5.3",
+    "esbuild": "0.14.38",
+    "esbuild-register": "3.3.2",
+    "eslint": "8.15.0",
+    "eslint-config-prettier": "8.5.0",
     "eslint-plugin-eslint-comments": "3.2.0",
-    "eslint-plugin-import": "2.25.4",
+    "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jest": "26.1.5",
     "eslint-plugin-prettier": "4.0.0",
     "eslint-plugin-simple-import-sort": "7.0.0",
     "eventemitter3": "4.0.7",
     "execa": "5.1.1",
     "glob": "8.0.1",
-    "globby": "11.0.4",
+    "globby": "11.1.0",
     "graphviz-mit": "0.0.9",
-    "husky": "8.0.0",
+    "husky": "8.0.1",
     "is-ci": "3.0.1",
     "jest-junit": "13.2.0",
-    "lint-staged": "12.3.4",
+    "lint-staged": "12.4.1",
     "node-fetch": "2.6.7",
     "p-map": "4.0.0",
     "p-reduce": "2.1.0",
-    "p-retry": "4.6.1",
+    "p-retry": "4.6.2",
     "path-browserify": "1.0.1",
-    "prettier": "2.5.1",
+    "prettier": "2.6.2",
     "redis": "3.1.2",
     "redis-lock": "0.1.4",
     "regenerator-runtime": "0.13.9",
-    "resolve": "1.21.0",
+    "resolve": "1.22.0",
     "safe-buffer": "5.2.1",
-    "semver": "7.3.5",
+    "semver": "7.3.7",
     "spdx-exceptions": "2.3.0",
     "spdx-license-ids": "3.0.11",
-    "staged-git-files": "1.2.0",
-    "ts-node": "10.4.0",
+    "staged-git-files": "1.3.0",
+    "ts-node": "10.7.0",
     "ts-toolbelt": "9.6.0",
     "tty-browserify": "0.0.1",
-    "typescript": "4.5.4",
+    "typescript": "4.6.4",
     "util": "0.12.4"
   },
   "lint-staged": {
