@@ -75,7 +75,7 @@
     "@prisma/sdk": "workspace:*",
     "@prisma/studio": "0.460.0",
     "@prisma/studio-server": "0.460.0",
-    "@swc/core": "1.2.141",
+    "@swc/core": "1.2.182",
     "@swc/jest": "0.2.21",
     "@types/debug": "4.1.7",
     "@types/fs-extra": "9.0.13",
@@ -84,13 +84,13 @@
     "@types/ws": "8.5.3",
     "chalk": "4.1.2",
     "checkpoint-client": "1.1.21",
-    "debug": "4.3.3",
-    "dotenv": "16.0.0",
-    "esbuild": "0.13.14",
+    "debug": "4.3.4",
+    "dotenv": "16.0.1",
+    "esbuild": "0.14.38",
     "execa": "5.1.1",
     "fast-deep-equal": "3.1.3",
-    "fs-extra": "10.0.0",
-    "fs-jetpack": "4.3.0",
+    "fs-extra": "10.1.0",
+    "fs-jetpack": "4.3.1",
     "get-port": "5.1.1",
     "global-dirs": "3.0.0",
     "is-installed-globally": "0.4.0",
@@ -108,7 +108,7 @@
     "strip-ansi": "6.0.1",
     "tempy": "1.0.1",
     "ts-pattern": "4.0.2",
-    "typescript": "4.5.4"
+    "typescript": "4.6.4"
   },
   "scripts": {
     "dev": "DEV=true node -r esbuild-register helpers/build.ts",
