@@ -62,8 +62,8 @@
   ],
   "devDependencies": {
     "@jest/test-sequencer": "28.1.0",
-    "@microsoft/api-extractor": "7.19.3",
-    "@opentelemetry/api": "1.0.3",
+    "@microsoft/api-extractor": "7.23.2",
+    "@opentelemetry/api": "1.1.0",
     "@prisma/debug": "workspace:*",
     "@prisma/engine-core": "workspace:*",
     "@prisma/engines": "3.15.0-3.55eef363a3b2e17691b1f98ed28da8b387f3bdb8",
@@ -72,7 +72,7 @@
     "@prisma/get-platform": "3.15.0-3.55eef363a3b2e17691b1f98ed28da8b387f3bdb8",
     "@prisma/migrate": "workspace:*",
     "@prisma/sdk": "workspace:*",
-    "@swc/core": "1.2.141",
+    "@swc/core": "1.2.182",
     "@swc/jest": "0.2.21",
     "@timsuchanek/copy": "1.4.5",
     "@types/debug": "4.1.7",
@@ -85,10 +85,10 @@
     "benchmark": "2.1.4",
     "chalk": "4.1.2",
     "decimal.js": "10.3.1",
-    "esbuild": "0.13.14",
+    "esbuild": "0.14.38",
     "execa": "5.1.1",
     "flat-map-polyfill": "0.3.8",
-    "fs-extra": "10.0.0",
+    "fs-extra": "10.1.0",
     "fs-monkey": "1.0.3",
     "get-own-enumerable-property-symbols": "3.0.2",
     "globby": "11.1.0",
@@ -102,8 +102,8 @@
     "lz-string": "1.4.4",
     "make-dir": "3.1.0",
     "mariadb": "3.0.0",
-    "mssql": "8.0.1",
-    "pg": "8.7.1",
+    "mssql": "8.1.0",
+    "pg": "8.7.3",
     "pkg-up": "3.1.0",
     "pluralize": "8.0.0",
     "replace-string": "3.1.0",
@@ -116,9 +116,9 @@
     "strip-ansi": "6.0.1",
     "strip-indent": "3.0.0",
     "ts-jest": "28.0.2",
-    "ts-node": "10.4.0",
-    "tsd": "0.19.1",
-    "typescript": "4.5.4"
+    "ts-node": "10.7.0",
+    "tsd": "0.20.0",
+    "typescript": "4.6.4"
   },
   "peerDependencies": {
     "prisma": "*"
