@@ -20,11 +20,11 @@
   "devDependencies": {
     "@types/jest": "27.5.0",
     "@types/node": "12.20.51",
-    "esbuild": "0.13.14",
+    "esbuild": "0.14.38",
     "jest": "28.1.0",
     "jest-junit": "13.2.0",
     "ts-jest": "28.0.2",
-    "typescript": "4.5.4"
+    "typescript": "4.6.4"
   },
   "scripts": {
     "dev": "DEV=true node -r esbuild-register helpers/build.ts",
@@ -38,8 +38,8 @@
   ],
   "dependencies": {
     "@types/debug": "4.1.7",
-    "ms": "2.1.3",
+    "debug": "4.3.4",
     "strip-ansi": "6.0.1"
   },
-  "sideEffects": true
+  "sideEffects": false
 }
