@@ -3,7 +3,6 @@
   "version": "0.0.0",
   "license": "Apache-2.0",
   "main": "dist/index.js",
-  "module": "esm/dist/index.mjs",
   "types": "dist/index.d.ts",
   "author": "Tim Suchanek <suchanek@prisma.io>",
   "homepage": "https://www.prisma.io",
@@ -19,14 +18,14 @@
   ],
   "bugs": "https://github.com/prisma/prisma/issues",
   "devDependencies": {
-    "@swc/core": "1.2.141",
+    "@swc/core": "1.2.182",
     "@swc/jest": "0.2.21",
     "@types/jest": "27.5.0",
     "@types/node": "16.11.34",
-    "esbuild": "0.13.14",
+    "esbuild": "0.14.38",
     "jest": "28.1.0",
     "jest-junit": "13.2.0",
-    "typescript": "4.5.4"
+    "typescript": "4.6.4"
   },
   "scripts": {
     "dev": "DEV=true node -r esbuild-register helpers/build.ts",
