@@ -2,7 +2,6 @@
   "name": "@prisma/generator-helper",
   "version": "0.0.0",
   "main": "dist/index.js",
-  "module": "esm/dist/index.mjs",
   "types": "dist/index.d.ts",
   "typings": "dist/index.d.ts",
   "license": "Apache-2.0",
@@ -32,15 +31,15 @@
     "prepublishOnly": "pnpm run build"
   },
   "devDependencies": {
-    "@swc/core": "1.2.141",
+    "@swc/core": "1.2.182",
     "@swc/jest": "0.2.21",
     "@types/jest": "27.5.0",
     "@types/node": "12.20.51",
-    "esbuild": "0.13.14",
+    "esbuild": "0.14.38",
     "jest": "28.1.0",
     "jest-junit": "13.2.0",
-    "ts-node": "10.4.0",
-    "typescript": "4.5.4"
+    "ts-node": "10.7.0",
+    "typescript": "4.6.4"
   },
   "files": [
     "README.md",
