@@ -3,7 +3,6 @@
   "version": "0.0.0",
   "private": true,
   "main": "dist/index.js",
-  "module": "esm/dist/index.mjs",
   "types": "dist/index.d.ts",
   "license": "Apache-2.0",
   "author": "Tim Suchanek <suchanek@prisma.io>",
@@ -15,7 +14,7 @@
   "devDependencies": {
     "@prisma/sdk": "workspace:*",
     "@sindresorhus/slugify": "1.1.2",
-    "@swc/core": "1.2.141",
+    "@swc/core": "1.2.182",
     "@swc/jest": "0.2.21",
     "@types/jest": "27.5.0",
     "@types/mssql": "7.1.5",
@@ -23,21 +22,21 @@
     "@types/pg": "8.6.5",
     "@types/sqlite3": "3.1.8",
     "decimal.js": "10.3.1",
-    "esbuild": "0.13.14",
+    "esbuild": "0.14.38",
     "execa": "5.1.1",
-    "fs-jetpack": "4.3.0",
+    "fs-jetpack": "4.3.1",
     "jest": "28.1.0",
     "jest-junit": "13.2.0",
     "mariadb": "3.0.0",
-    "mssql": "8.0.1",
-    "pg": "8.7.1",
+    "mssql": "8.1.0",
+    "pg": "8.7.3",
     "replace-string": "3.1.0",
-    "sqlite-async": "1.1.2",
+    "sqlite-async": "1.1.3",
     "string-hash": "1.1.3",
     "strip-ansi": "6.0.1",
     "tempy": "1.0.1",
-    "ts-node": "10.4.0",
-    "typescript": "4.5.4",
+    "ts-node": "10.7.0",
+    "typescript": "4.6.4",
     "verror": "1.10.1"
   },
   "scripts": {
