@@ -2,7 +2,6 @@
   "name": "@prisma/sdk",
   "version": "0.0.0",
   "main": "dist/index.js",
-  "module": "esm/dist/index.mjs",
   "types": "dist/index.d.ts",
   "repository": {
     "type": "git",
@@ -30,19 +29,19 @@
     "scripts"
   ],
   "devDependencies": {
-    "@swc/core": "1.2.141",
+    "@swc/core": "1.2.182",
     "@swc/jest": "0.2.21",
     "@types/jest": "27.5.0",
     "@types/node": "12.20.51",
     "@types/resolve": "1.20.2",
     "@types/shell-quote": "1.7.1",
     "@types/tar": "6.1.1",
-    "esbuild": "0.13.14",
+    "esbuild": "0.14.38",
     "jest": "28.1.0",
     "jest-junit": "13.2.0",
     "mock-stdin": "1.0.0",
-    "ts-node": "10.4.0",
-    "typescript": "4.5.4",
+    "ts-node": "10.7.0",
+    "typescript": "4.6.4",
     "yarn": "1.22.18"
   },
   "dependencies": {
