@@ -13,7 +13,7 @@
   "devDependencies": {
     "@types/node": "17.0.21",
     "prisma": "../../packages/cli",
-    "ts-node": "10.4.0",
-    "typescript": "4.5.5"
+    "ts-node": "10.7.0",
+    "typescript": "4.6.4"
   }
 }
