@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-this-alias */
 // Copyright (C) 2011-2015 John Hewson
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -18,16 +19,18 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 // IN THE SOFTWARE.
 
+// @ts-ignore
+
 import stream from 'stream'
 import util from 'util'
 
-// convinience API
+// convenience API
 export default function byline(readStream, options?: any) {
-  return module.exports.createStream(readStream, options)
+  return createStream(readStream, options)
 }
 
 // basic API
-module.exports.createStream = function (readStream, options) {
+export function createStream(readStream, options) {
   if (readStream) {
     return createLineStream(readStream, options)
   } else {
@@ -51,21 +54,19 @@ export function createLineStream(readStream, options) {
 // using the new node v0.10 "streams2" API
 //
 
-module.exports.LineStream = LineStream
-
-function LineStream(this, options) {
+export function LineStream(this: any, options) {
   stream.Transform.call(this, options)
   options = options || {}
 
   // use objectMode to stop the output from being buffered
-  // which re-concatanates the lines, just without newlines.
+  // which re-concatenates the lines, just without newlines.
   this._readableState.objectMode = true
   this._lineBuffer = []
   this._keepEmptyLines = options.keepEmptyLines || false
   this._lastChunkEndedWithCR = false
 
   // take the source's encoding if we don't have one
-  this.on('pipe', function (this, src) {
+  this.on('pipe', function (this: any, src) {
     if (!this.encoding) {
       // but we can't do this for old-style streams
       if (src instanceof stream.Readable) {
@@ -115,7 +116,6 @@ LineStream.prototype._pushBuffer = function (encoding, keep, done) {
     if (this._keepEmptyLines || line.length > 0) {
       if (!this.push(this._reencode(line, encoding))) {
         // when the high-water mark is reached, defer pushes until the next tick
-        // eslint-disable-next-line @typescript-eslint/no-this-alias
         const self = this
         setImmediate(function () {
           self._pushBuffer(encoding, keep, done)
