@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-this-alias */
 // Copyright (C) 2011-2015 John Hewson
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -18,16 +19,18 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 // IN THE SOFTWARE.
 
+// @ts-ignore
+
 import stream from 'stream'
 import util from 'util'
 
-// convinience API
+// convenience API
 export default function byline(readStream, options?: any) {
-  return module.exports.createStream(readStream, options)
+  return createStream(readStream, options)
 }
 
 // basic API
-module.exports.createStream = function (readStream, options) {
+export function createStream(readStream, options) {
   if (readStream) {
     return createLineStream(readStream, options)
   } else {
@@ -51,14 +54,12 @@ export function createLineStream(readStream, options) {
 // using the new node v0.10 "streams2" API
 //
 
-module.exports.LineStream = LineStream
-
-function LineStream(this: any, options) {
+export function LineStream(this: any, options) {
   stream.Transform.call(this, options)
   options = options || {}
 
   // use objectMode to stop the output from being buffered
-  // which re-concatanates the lines, just without newlines.
+  // which re-concatenates the lines, just without newlines.
   this._readableState.objectMode = true
   this._lineBuffer = []
   this._keepEmptyLines = options.keepEmptyLines || false
