@@ -42,20 +42,15 @@ const proxyBuildConfig: BuildOptions = {
     'define.amd': 'false',
   },
   plugins: [
-    fillPlugin(
-      {
-        // TODO no tree shaking on wrapper pkgs
-        '@prisma/get-platform': { contents: '' },
-        // removes un-needed code out of `chalk`
-        'supports-color': { contents: '' },
-        // these can not be exported any longer
-        './warnEnvConflicts': { contents: '' },
-        './utils/find': { contents: '' },
-      },
-      // we only trigger it on the first step (esm)
-      // because that is where tree-shaking happens
-      (options) => options.format === 'esm',
-    ),
+    fillPlugin({
+      // TODO no tree shaking on wrapper pkgs
+      '@prisma/get-platform': { contents: '' },
+      // removes un-needed code out of `chalk`
+      'supports-color': { contents: '' },
+      // these can not be exported any longer
+      './warnEnvConflicts': { contents: '' },
+      './utils/find': { contents: '' },
+    }),
   ],
   logLevel: 'error',
 }
@@ -89,6 +84,11 @@ function bundleTypeDefinitions(filename: string, outfile: string) {
       ],
       compiler: {
         tsconfigFilePath: 'tsconfig.build.json',
+        overrideTsconfig: {
+          compilerOptions: {
+            paths: {}, // bug with api extract + paths
+          },
+        },
       },
       dtsRollup: {
         enabled: true,
@@ -117,8 +117,8 @@ function bundleTypeDefinitions(filename: string, outfile: string) {
 
 void build([generatorBuildConfig, runtimeBuildConfig, browserBuildConfig, proxyBuildConfig]).then(() => {
   if (process.env.DEV !== 'true') {
-    bundleTypeDefinitions('declaration/runtime/index', 'runtime/index')
-    bundleTypeDefinitions('declaration/runtime/index', 'runtime/proxy')
-    bundleTypeDefinitions('declaration/runtime/index-browser', 'runtime/index-browser')
+    bundleTypeDefinitions('declaration/client/src/runtime/index', 'runtime/index')
+    bundleTypeDefinitions('declaration/client/src/runtime/index', 'runtime/proxy')
+    bundleTypeDefinitions('declaration/client/src/runtime/index-browser', 'runtime/index-browser')
   }
 })
