@@ -3,17 +3,18 @@ import stripAnsi from 'strip-ansi'
 import { removeISODate, sanitizeTestLogs } from '../util'
 
 describe('debug', () => {
-  process.env.DEBUG = 'my-namespace'
-  const DebugLib = require('../')
-  const Debug = DebugLib.Debug
-  const getLogs = DebugLib.getLogs
+  test('env vars work as expected', async () => {
+    process.env.DEBUG = 'my-namespace'
 
-  test('env vars work as expected', () => {
+    const { Debug, getLogs } = await import('../index')
+
     const debug = Debug('my-namespace')
     const logs: string[] = []
+
     debug.log = (...args) => {
-      logs.push(stripAnsi(args[0]).trimStart())
+      logs.push(stripAnsi(`${args[0]}${args[1]}`).trim())
     }
+
     debug('Does it even log?')
     debug('I dont know')
 
