@@ -1,34 +1,58 @@
-/// <reference types="debug" />
-import DebugNode from './node'
-
-const cache: any[] = []
+import debug from 'debug'
 
 const MAX_LOGS = 100
 
-export default function Debug(namespace: string): debug.Debugger {
-  const debug: debug.Debugger = DebugNode(namespace, (...args) => {
-    cache.push(args)
-    // keeping 100 logs is just a heuristic. The real truncating comes later
-    if (cache.length > MAX_LOGS) {
-      cache.shift()
+const debugArgsHistory: any[] = []
+
+/**
+ * Wrapper on top of the original `Debug` to keep a history of the all last
+ * {@link MAX_LOGS}. This is then used by {@link getLogs} to generate an error
+ * report url (forGitHub) in the case where the something has crashed.
+ * @param namespace
+ * @returns
+ */
+function debugCall(namespace: string) {
+  const debugNamespace = debug(namespace)
+
+  // we take over the `debugNamespace` function
+  const call = Object.assign((...args: any[]) => {
+    // debug only calls log if you implement it
+    debugNamespace.log = (call as any).log
+
+    // we push the args to our history of args
+    if (args.length !== 0) {
+      debugArgsHistory.push([namespace, ...args])
     }
-  })
 
-  return debug
+    // if it is too big, then we remove some
+    if (debugArgsHistory.length > MAX_LOGS) {
+      debugArgsHistory.shift()
+    }
+
+    // we apply the function with no format
+    return debugNamespace('', ...args)
+  }, debugNamespace)
+
+  return call as debug.Debugger
 }
-export { Debug }
 
-Debug.enable = (namespace: string): void => {
-  DebugNode.enable(namespace)
-}
+/**
+ * This essentially mimics the original `debug` api. It is a debug function call
+ * that has utility properties on it. We provide our custom {@link debugCall},
+ * and expose the original original api as-is.
+ */
+const Debug = Object.assign(debugCall, debug)
 
-Debug.enabled = (namespace: string): boolean => DebugNode.enabled(namespace)
-
-// https://stackoverflow.com/questions/417142/what-is-the-maximum-length-of-a-url-in-different-browsers
-// we need some space for other characters, so we go for 30k here
+/**
+ * We can get the logs for all the last {@link MAX_LOGS} ${@link debugCall} that
+ * have happened in the different packages. Useful to generate error report links.
+ * @see https://stackoverflow.com/questions/417142/what-is-the-maximum-length-of-a-url-in-different-browsers
+ * @param numChars
+ * @returns
+ */
 export function getLogs(numChars = 7500): string {
   // flatmap on text level
-  const output = cache
+  const output = debugArgsHistory
     .map((c) =>
       c
         .map((item) => {
@@ -38,7 +62,7 @@ export function getLogs(numChars = 7500): string {
 
           return JSON.stringify(item)
         })
-        .join('  '),
+        .join(' '),
     )
     .join('\n')
 
@@ -48,3 +72,6 @@ export function getLogs(numChars = 7500): string {
 
   return output.slice(-numChars)
 }
+
+export { Debug }
+export default Debug
